+++ MinimOsd_Extra/eeprom.h
@@ -190,6 +190,8 @@ struct Settings {
 
 //0x45
     byte pwm_mode; // 0-pwm 1-in/off
+    byte hw_version; // written by firmware for CT
+    
 };
 
 union _Settings {